// lva_project/ClientApp/src/app/models/lva.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export class Lva {
    lvaId: number = 0;
    lvaName: string = "";
    lvaTeacher: string = "";
    lvaNumber: number = 0;
    lvaType: number = 0;
    lvaRoom: string = "";
    lvaInstitute: string = "";
    lvaExam: Date | null = null;
    lvaCreatedOn: Date = new Date();
    lvaEcts: number = 0;
    lvaGrade: number = 0;
    semId: number = 0;
}